+++ src/main/java/ar/edu/utn/frba/dds/models/entities/cuestionario/Respuesta.java
@@ -13,23 +13,20 @@ public class Respuesta {
 
     @Id
     @GeneratedValue ( strategy = GenerationType.IDENTITY)
-    private Integer id_Respuesta;
+    private Long id_Respuesta;
     @ManyToOne
     @JoinColumn(name = "id_Pregunta")
     private Pregunta pregunta;
-
 
     @ManyToOne
     @JoinColumn(name = "id_CuestionarioRespondido")
     private CuestionarioRespondido cuestionarioRespondido;
 
-    //private List<Opcion> opcion;
     @Column(name = "respuestaAbierta", columnDefinition = "TEXT")
     private String respuestaAbierta;
 
     public Respuesta(Pregunta pregunta,String r123espuestaAbierta){
         this.pregunta = pregunta;
-        //this.opcion = new ArrayList<Opcion>();
         this.respuestaAbierta = respuestaAbierta;
     }
 
@@ -38,7 +35,5 @@ public class Respuesta {
     public boolean esObligatoria(){
         return pregunta.esObligatoria();
     }
-
-    //public void agregarOpcion(Opcion opcion){this.opcion.add(opcion);}
 
 }
